Added: Two overloaded Integrand::finalizeElement methods for flexibility, and some more element data in the FiniteElement class.
git-svn-id: http://svn.sintef.no/trondheim/IFEM/trunk@2866 e10b68d5-8a6e-419e-a041-bce267b0401d

+++ src/ASM/ASMs1D.h
@@ -303,8 +303,8 @@ protected:
 
   //! \brief Returns nodal rotation matrices for an element, if any.
   //! \param[out] T Array of nodal rotation matrices
-  //! \param[in] iel Element index
-  bool getElementNodalRotations(TensorVec& T, int iel) const;
+  //! \param[in] iel 0-based element index
+  bool getElementNodalRotations(TensorVec& T, size_t iel) const;
 
 public:
   //! \brief Auxilliary function for computation of basis function indices.
